Move and rename MainCommand -> KaboomCommand
